SCRIPTS/jira/requirements.txt
atlassian-python-api==3.10.0
requests==2.22.0
pyyaml==5.4